refactor: move CURIE class to Helpers folder
