+++ src/API/REST/V3/Routes/Donors/DonorController.php
@@ -2,7 +2,7 @@
 
 namespace Give\API\REST\V3\Routes\Donors;
 
-use Give\API\REST\V3\Routes\CURIE;
+use Give\API\REST\V3\Routes\Helpers\CURIE;
 use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorAnonymousMode;
 use Give\API\REST\V3\Routes\Donors\ValueObjects\DonorRoute;
 use Give\Donors\DonorsQuery;
